third_party/blink/web_tests/fast/dom/HTMLLinkElement/disabled-attribute-expected.txt

This test checks HTMLLinkElement disabled conforms to HTML5 spec for the bug id #45425 .

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS test.disabled is false
PASS test.disabled is true
PASS test.sheet is null


PASS test_nostyle.sheet is null
PASS test_nostyle.disabled is false
PASS test_nostyle.disabled is true
PASS successfullyParsed is true

TEST COMPLETE
